requirements.txt
python-dotenv
fastapi
uvicorn[standard]
pydantic
SQLAlchemy
psycopg2-binary
email-validator